Comment code refactored.
