+++ src/main/java/ImageHoster/repository/ImageRepository.java
@@ -113,16 +113,4 @@ public class ImageRepository {
         }
     }
 
-    public void postComment(Comment comment) {
-        EntityManager em = emf.createEntityManager();
-        EntityTransaction transaction = em.getTransaction();
-
-        try {
-            transaction.begin();
-            em.persist(comment);
-            transaction.commit();
-        } catch (Exception e) {
-            transaction.rollback();
-        }
-    }
 }
